requirements.txt
SQLAlchemy
pillow
